# terraform_registry_api/terraform_module_registry_api/swagger.yml
swagger: "2.0"
info:
  description: This is the swagger file that goes with our server code
  version: "1.0.0"
  title: Swagger Rest Article
consumes:
  - application/json
produces:
  - application/json

basePath: /v1/modules

# Paths supported by the server application
paths:
  /{namespace}/{name}/{provider}/versions:
    get:
      operationId: terraform_registry_api.terraform_module_registry_api.api.list_versions
      tags:
        - Modules
        - Base
      summary: List Module Versions
      description: Listing the module versions
      parameters:
          - name: namespace
            in: path
            description: Module namespace
            type: string
            required: True
          - name: name
            in: path
            description: Name of the module
            type: string
            required: True
          - name: provider
            in: path
            description: the system target
            type: string
            required: True
      responses:
        200:
          description: Successfully list version of module
          schema:
            properties:
              modules:
                type: string
  /{namespace}/{name}/{provider}/{version}/download:
    get:
      operationId: terraform_registry_api.terraform_module_registry_api.api.download_version
      tags:
        - Modules
        - Base
      summary: Download a single Module Version
      description: Download a version
      parameters:
        - name: namespace
          in: path
          description: Module namespace
          type: string
          required: True
        - name: name
          in: path
          description: Name of the module
          type: string
          required: True
        - name: provider
          in: path
          description: the system target
          type: string
          required: True
        - name: version
          in: path
          description: The version of the module
          type: string
          required: True
      responses:
        204:
          description: Empty Response with download header
  /{namespace}:
    get:
      operationId: terraform_registry_api.terraform_module_registry_api.api.list_modules
      tags:
        - Modules
        - Extended
      summary: List Modules
      description: Listing the modules
      parameters:
        - name: namespace
          in: path
          description: Module namespace
          type: string
          required: True
        - name: offset
          in: query
          description: Module namespace
          type: integer
          required: False
        - name: limit
          in: query
          description: Module namespace
          type: integer
          required: False
        - name: provider
          in: query
          description: Limits modules to a specific provider.
          type: string
          required: False
        - name: verified
          in: query
          description: Limits modules to verified modules.
          type: boolean
          required: False
      responses:
        200:
          description: Successfully list version of module
          schema:
            properties:
              modules:
                type: string
  /:
    get:
      operationId: terraform_registry_api.terraform_module_registry_api.api.list_all_modules
      tags:
        - Modules
        - Extended
      summary: List All Modules
      description: Listing all of the modules
      parameters:
        - name: offset
          in: query
          description: Module namespace
          type: integer
          required: False
        - name: limit
          in: query
          description: Module namespace
          type: integer
          required: False
        - name: provider
          in: query
          description: Limits modules to a specific provider.
          type: string
          required: False
        - name: verified
          in: query
          description: Limits modules to verified modules.
          type: boolean
          required: False
      responses:
        200:
          description: Successfully list version of module
          schema:
            properties:
              modules:
                type: string
  /search:
    get:
      operationId: terraform_registry_api.terraform_module_registry_api.api.search_modules
      tags:
        - Modules
        - Extended
      summary: Search Modules
      description: Listing the modules
      parameters:
        - name: q
          in: query
          description: The search string.
          type: string
          required: True
        - name: offset
          in: query
          description: Module namespace
          type: integer
          required: False
        - name: limit
          in: query
          description: Module namespace
          type: integer
          required: False
        - name: provider
          in: query
          description: Limits modules to a specific provider.
          type: string
          required: False
        - name: verified
          in: query
          description: Limits modules to verified modules.
          type: boolean
          required: False
        - name: namespace
          in: query
          description: Limits results to a specific namespace.
          type: string
          required: False
      responses:
        200:
          description: Successfully list version of module
          schema:
            properties:
              modules:
                type: string
  /{namespace}/{name}:
    get:
      operationId: terraform_registry_api.terraform_module_registry_api.api.get_latest_for_all_providers
      tags:
        - Modules
        - Extended
      summary: List Latest Version of Module for All Providers
      description: This endpoint returns the latest version of each provider for a module.
      parameters:
        - name: namespace
          in: path
          description: Module namespace
          type: string
          required: True
        - name: name
          in: path
          description: Name of the module
          type: string
          required: True
        - name: offset
          in: query
          description: Module namespace
          type: integer
          required: False
        - name: limit
          in: query
          description: Module namespace
          type: integer
          required: False
      responses:
        200:
          description: Successfully list version of module
          schema:
            properties:
              modules:
                type: string
  /{namespace}/{name}/{provider}:
    get:
      operationId: terraform_registry_api.terraform_module_registry_api.api.get_latest_for_provider
      tags:
        - Modules
        - Extended
      summary: Latest Version for a Specific Module Provider
      description: This endpoint returns the latest version of a module for a single provider.
      parameters:
        - name: namespace
          in: path
          description: Module namespace
          type: string
          required: True
        - name: name
          in: path
          description: Name of the module
          type: string
          required: True
        - name: provider
          in: path
          description: the system target
          type: string
          required: True
      responses:
        200:
          description: Successfully list version of module
          schema:
            properties:
              modules:
                type: string
  /{namespace}/{name}/{provider}/{version}:
    get:
      operationId: terraform_registry_api.terraform_module_registry_api.api.get_module
      tags:
        - Modules
        - Extended
      summary: Get a Specific Module
      description: This endpoint returns the specified version of a module for a single provider.
      parameters:
        - name: namespace
          in: path
          description: Module namespace
          type: string
          required: True
        - name: name
          in: path
          description: Name of the module
          type: string
          required: True
        - name: provider
          in: path
          description: the system target
          type: string
          required: True
        - name: version
          in: path
          description: The version of the module
          type: string
          required: True
      responses:
        200:
          description: Successfully list version of module
          schema:
            properties:
              modules:
                type: string
  /{namespace}/{name}/{provider}/download:
    get:
      operationId: terraform_registry_api.terraform_module_registry_api.api.download_latest
      tags:
        - Modules
        - Extended
      summary: Download the Latest Version of a Module
      description: |
        This endpoint downloads the latest version of a module for a single provider.

        It returns a 302 redirect whose Location header redirects the client to the download endpoint (above) for the latest version.
      parameters:
        - name: namespace
          in: path
          description: Module namespace
          type: string
          required: True
        - name: name
          in: path
          description: Name of the module
          type: string
          required: True
        - name: provider
          in: path
          description: the system target
          type: string
          required: True
      responses:
        302:
          description: Successfully list version of module
  /{namespace}/{name}/{provider}/{version}/upload:
    post:
      operationId: terraform_registry_api.terraform_module_registry_api.api.upload_version
      tags:
        - Modules
        - Base
      summary: Upload a single Module Version
      description: Upload a version
      parameters:
        - name: namespace
          in: path
          description: Module namespace
          type: string
          required: True
        - name: name
          in: path
          description: Name of the module
          type: string
          required: True
        - name: provider
          in: path
          description: the system target
          type: string
          required: True
        - name: version
          in: path
          description: The version of the module
          type: string
          required: True
      responses:
        200:
          description: Empty Response with download header